requirements.txt
fastapi
uvicorn
python-multipart
GitPython
openai
apscheduler
